Fix session cleanup bug
The function swi_mangoh_data_router_removeAllUpdateHandlersForSession
was iterating over database items incorrectly which meant that the
code was executing as though there was nothing to clean up when there
was data to clean and executing as though there was something to
clean up when there was nothing.  In the latter case, this caused an
assert in the data router.  This bug was also masking another bug in
the ListRemoveGeneral() helper function.  In addition, the severity of
the "session not found" messages has been incresed to error because
reads/writes of values should never occur without a session.

Resolves: MANGOH-4
Change-Id: If50b3ef1b8936e6ee81ee5de01ef9a5fd341da10